Quiz/Reviewer
The Quiz Generator is now part of the review.html. The + button when in the review.html page is now brings up the quizGenerator page. Questions, answers, and choices are now stored in the local storage so that it can be fetched from the different parts of the app. There is now a quiz page in the review.html page and when buttons(eg. Quiz 1) are for example clicked, it will be directed to the quiz page, the generated quiz will be loaded for the users use.